Added patch from Rune Glerup:
OSGi (see http://www.osgi.org) is a standard that enables highly modular Java programs by specifying a module concept called a bundle. A bundle is a jar that explicitly offers classes and services, and declares its dependencies on other bundles' resources and services. At runtime each bundle has its own classloader that enforces the import / export declarations by providing access to the bundle's declared dependencies.

This scheme sometimes interacts in unexpected ways with libraries that use runtime code generation. A technique called bridge classloading can be used to make it work in an OSGi setting. I have implemented the technique as a patch to Dirmi, and I would like to offer the patch to you. The bridge classloader will have no effect when running outside an OSGi environment.

Including this patch and adding a few OSGi-related properties to the MANIFEST.MF file would turn the Dirmi jar into a valid OSGi bundle. I would be happy to contribute the properties that needs to be written in the MANIFEST.MF file if you are interested.

== Bridge classloading explained ==

When using Dirmi in an OSGi application there are two main participating bundles:

 1. The Dirmi library bundle
 2. The application bundle(s) containing the remote object interfaces

When generating stubs for the remote objects, Dirmi will generally use the classloader of the application bundle containing the remote interface. Since the application bundle should not depend on the Dirmi bundle, its classloader will not be able to load the Dirmi classes needed in generating the stubs.

The solution is to generate a bridge classloader for each application bundle classloader that simply delegates to the application bundle classloader, except when loading classes that starts with "org.cojen.dirmi" in which case it delegates to the Dirmi bundle classloader.

The technique is described in more detail in this article: http://www.infoq.com/articles/code-generation-with-osgi

Another example of its use is in the Google Guice library: http://code.google.com/p/google-guice/source/browse/trunk/core/src/com/google/inject/internal/BytecodeGen.java

I hope you will consider including this patch, and I would be happy to work with you on adapting the solution if there are parts of it, you are not happy with.